fix date-time format validation
The current regex to validate date-time doesn't capture all the
invalid dates. Also the spec mandates proper day validation, that is,
february can't have 29 on non leap year etc which can't be validated
via regex.

This commit uses the DateTime module from standard library to do the
validation. Given that the function is available since 1.4, support
for 1.3 is dropped.