Fix bug when validating schemas that specify an explicit metaschema
Schema.assert_valid_schema/1 was checking whether an input schema was one of the JSON Schema draft metaschemas, and if so, automatically marking it as valid.  The problem, though, is that it was checking the $schema field instead of the $id field.  This meant that any schema which specified that it's supposed to be checked against a JSON Schema draft would automatically pass even if it wasn't valid.